typo in ci yml, licenses, rename module
