+++ .github/workflows/ci.yml
@@ -20,7 +20,7 @@ env:
   CARGO_HTTP_MULTIPLEXING: false
   CARGO_TERM_COLOR: always
   PROTOC: protoc
-  TERM: unkown
+  TERM: unknown
 
 jobs:
   clippy:
