+++ .license.ignore
@@ -4,5 +4,6 @@
 ./comms/tari_comms_logging/src/schema.rs
 ./dan_layer/storage_sqlite/src/global/schema.rs
 ./dan_layer/storage_sqlite/src/schema.rs
+./dan_layer/state_store_sqlite/src/schema.rs
 ./dan_layer/wallet/storage_sqlite/src/schema.rs
 ./scripts/env_sample
