+++ dan_layer/state_store_sqlite/src/lib.rs
@@ -1,7 +1,9 @@
-mod deser;
+//   Copyright 2023 The Tari Project
+//   SPDX-License-Identifier: BSD-3-Clause
 mod error;
 mod reader;
 mod schema;
+mod serialization;
 mod sql_models;
 mod sqlite_transaction;
 mod store;
