+++ dan_layer/state_store_sqlite/src/serialization.rs
