+++ dan_layer/state_store_sqlite/src/sql_models/block.rs
@@ -6,7 +6,7 @@ use tari_dan_common_types::{Epoch, NodeHeight};
 use tari_dan_storage::{consensus_models, StorageError};
 use time::PrimitiveDateTime;
 
-use crate::deser::{deserialize_hex_try_from, deserialize_json};
+use crate::serialization::{deserialize_hex_try_from, deserialize_json};
 
 #[derive(Debug, Clone, Queryable)]
 pub struct Block {
